+++ examples/fc_networks.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 ###
-# (C) Copyright [2019] Hewlett Packard Enterprise Development LP
+# (C) Copyright [2020] Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -20,13 +20,14 @@ from hpOneView.oneview_client import OneViewClient
 from hpOneView.exceptions import HPOneViewException
 from config_loader import try_load_from_file
 
+# To run this example fill the ip and the credentials below or use a configuration file
 config = {
     "ip": "<oneview_ip>",
     "credentials": {
         "userName": "<username>",
         "password": "<password>"
     },
-    "api_version": 1200
+    "api_version": "<api_version>"
 }
 
 options = {
@@ -35,6 +36,14 @@ options = {
     "autoLoginRedistribution": True,
     "fabricType": "FabricAttach",
     "linkStabilityTime": 30,
+}
+
+options_bulk_delete = {
+    "networkUris": [
+        "/rest/fc-networks/e2f0031b-52bd-4223-9ac1-d91cb519d548",
+        "/rest/fc-networks/f2f0031b-52bd-4223-9ac1-d91cb519d549",
+        "/rest/fc-networks/02f0031b-52bd-4223-9ac1-d91cb519d54a"
+    ]
 }
 
 # Scope name to perform the patch operation
@@ -89,11 +98,11 @@ fc_network.update(data=data_to_update)
 print("\nUpdated fc-network '%s' successfully.\n  uri = '%s'" % (fc_network.data['name'], fc_network.data['uri']))
 print("  with attribute {'autoLoginRedistribution': %s}" % fc_network.data['autoLoginRedistribution'])
 
-# Adds fcnetwork to scope defined
-print("\nGet scope then add the network to it")
-scope = oneview_client.scopes.get_by_name(scope_name)
-print(scope['uri'])
-if scope:
+# Adds fc-network to scope defined only for V300 and V500
+if scope_name and 300 <= oneview_client.api_version <= 500:
+    print("\nGet scope then add the network to it")
+    scope = oneview_client.scopes.get_by_name(scope_name)
+    print(scope['uri'])
     try:
         fc_with_scope = fc_network.patch('replace',
                                          '/scopeUris',
@@ -105,3 +114,9 @@ if scope:
 # Delete the created network
 fc_network.delete()
 print("\nSuccessfully deleted fc-network")
+
+# Delete bulk fcnetworks
+if oneview_client.api_version >= 1600:
+    print("\nDelete bulk fcnetworks")
+    fc_network.delete_bulk(options_bulk_delete)
+    print("Successfully deleted bulk fcnetworks")