+++ examples/fcoe_networks.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 ###
-# (C) Copyright [2019] Hewlett Packard Enterprise Development LP
+# (C) Copyright [2020] Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -19,18 +19,28 @@ from pprint import pprint
 from hpOneView.oneview_client import OneViewClient
 from config_loader import try_load_from_file
 
+# To run this example fill the ip and the credentials below or use a configuration file
 config = {
     "ip": "<oneview_ip>",
     "credentials": {
         "userName": "<username>",
         "password": "<password>"
-    }
+    },
+    "api_version": "<api_version>"
 }
 
 options = {
     "name": "OneViewSDK Test FCoE Network",
     "vlanId": "201",
     "connectionTemplateUri": None,
+}
+
+options_bulk_delete = {
+    "networkUris": [
+        "/rest/fcoe-networks/e2f0031b-52bd-4223-9ac1-d91cb519d548",
+        "/rest/fcoe-networks/f2f0031b-52bd-4223-9ac1-d91cb519d549",
+        "/rest/fcoe-networks/02f0031b-52bd-4223-9ac1-d91cb519d54a"
+    ]
 }
 
 # Scope name to perform the patch operation
@@ -83,8 +93,8 @@ print("\nGet a fcoe-network by uri")
 fcoe_nets_by_uri = fcoe_networks.get_by_uri(fcoe_network.data['uri'])
 pprint(fcoe_nets_by_uri.data)
 
-# Adds FCOE network to scope defined
-if scope_name:
+# Adds FCOE network to scope defined only for V300 and V500
+if scope_name and 300 <= oneview_client.api_version <= 500:
     print("\nGet scope then add the network to it")
     scope = oneview_client.scopes.get_by_name(scope_name)
     fcoe_with_scope = fcoe_network.patch('replace',
@@ -95,3 +105,9 @@ if scope_name:
 # Delete the created network
 fcoe_network.delete()
 print("\nSuccessfully deleted fcoe-network")
+
+# Delete bulk fcoe networks
+if oneview_client.api_version >= 1600:
+    print("\nDelete bulk fcoe networks")
+    fcoe_network.delete_bulk(options_bulk_delete)
+    print("Successfully deleted bulk fcoe networks")